Task_2 assign by the CodSoft during their one month virtual internship program.
This is my second task, in which i tried to build a model for IRIS flower classification with using iris classification dataset provided by the CodSoft itself.

Till now i have performed the Exploratory data analysis on the dataset where i found the many approaches need to be applied for better training of the model such as Feature Engineering and many more.
My next step is apply those techniques and train the model as soon as possible.